refactor: refactor csvparsing into an object
