+++ server/src/test/kotlin/fi/oph/kitu/csvparsing/CsvParsingTest.kt
@@ -19,12 +19,14 @@ import kotlin.test.assertNull
 class CsvParsingTest {
     @Test
     fun `test yki suoritukset parsing`() {
-        val event = LoggerFactory.getLogger(CsvArgs::class.java).atInfo()
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event)
         val csv =
             """
             "1.2.246.562.24.20281155246","010180-9026","N","Öhmana-Testi","Ranja Testi","EST","Testikuja 5","40100","Testilä","testi@testi.fi",183424,2024-10-30T13:53:56Z,2024-09-01,"fin","YT","1.2.246.562.10.14893989377","Jyväskylän yliopisto, Soveltavan kielentutkimuksen keskus",2024-11-14,5,5,,5,5,,,,0,0,,
             """.trimIndent()
-        val suoritus = csv.asCsv<YkiSuoritusCsv>(CsvArgs(event = event))[0]
+        val suoritus = parser.convertCsvToData<YkiSuoritusCsv>(csv).first()
+
         val datePattern = "yyyy-MM-dd"
         val dateFormatter = DateTimeFormatter.ofPattern(datePattern)
         assertEquals(Oid("1.2.246.562.24.20281155246"), suoritus.suorittajanOID)
@@ -61,31 +63,37 @@ class CsvParsingTest {
 
     @Test
     fun `test legacy language code 10 parsing`() {
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event)
         val arvioijaCsv =
             """
             "1.2.246.562.24.24941612410","010180-922U","Torvinen-Testi","Anniina Testi","anniina.testi@yki.fi","Testiosoite 7357","00100","HELSINKI",1994-08-01,2019-06-29,2024-06-29,0,0,"10","PT+KT"
             """.trimIndent()
-        val arvioija = arvioijaCsv.asCsv<SolkiArvioijaResponse>()[0]
+        val arvioija = parser.convertCsvToData<SolkiArvioijaResponse>(arvioijaCsv)[0]
         assertEquals(Tutkintokieli.SWE10, arvioija.kieli)
     }
 
     @Test
     fun `test legacy language code 11 parsing`() {
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event)
         val arvioijaCsv =
             """
             "1.2.246.562.24.24941612410","010180-922U","Torvinen-Testi","Anniina Testi","anniina.testi@yki.fi","Testiosoite 7357","00100","HELSINKI",1994-08-01,2019-06-29,2024-06-29,0,0,"11","PT+KT"
             """.trimIndent()
-        val arvioija = arvioijaCsv.asCsv<SolkiArvioijaResponse>()[0]
+        val arvioija = parser.convertCsvToData<SolkiArvioijaResponse>(arvioijaCsv)[0]
         assertEquals(Tutkintokieli.ENG11, arvioija.kieli)
     }
 
     @Test
     fun `test legacy language code 12 parsing`() {
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event)
         val arvioijaCsv =
             """
             "1.2.246.562.24.24941612410","010180-922U","Torvinen-Testi","Anniina Testi","anniina.testi@yki.fi","Testiosoite 7357","00100","HELSINKI",1994-08-01,2019-06-29,2024-06-29,0,0,"12","PT+KT"
             """.trimIndent()
-        val arvioija = arvioijaCsv.asCsv<SolkiArvioijaResponse>()[0]
+        val arvioija = parser.convertCsvToData<SolkiArvioijaResponse>(arvioijaCsv)[0]
         assertEquals(Tutkintokieli.ENG12, arvioija.kieli)
     }
 
@@ -93,7 +101,8 @@ class CsvParsingTest {
     fun `test writing csv`() {
         val datePattern = "yyyy-MM-dd"
         val dateFormatter = DateTimeFormatter.ofPattern(datePattern)
-        val event = LoggerFactory.getLogger(CsvArgs::class.java).atInfo()
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event, useHeader = true)
         val entity =
             YkiSuoritusEntity(
                 id = null,
@@ -130,7 +139,7 @@ class CsvParsingTest {
             )
         val writable = listOf(entity.toYkiSuoritusCsv())
         val outputStream = ByteArrayOutputStream()
-        writable.writeAsCsv(outputStream, CsvArgs(useHeader = true, event = event))
+        parser.streamDataAsCsv(outputStream, writable)
         val expectedCsv =
             """
             suorittajanOID,hetu,sukupuoli,sukunimi,etunimet,kansalaisuus,katuosoite,postinumero,postitoimipaikka,email,suoritusID,lastModified,tutkintopaiva,tutkintokieli,tutkintotaso,jarjestajanOID,jarjestajanNimi,arviointipaiva,tekstinYmmartaminen,kirjoittaminen,rakenteetJaSanasto,puheenYmmartaminen,puhuminen,yleisarvosana,"tarkistusarvioinninSaapumisPvm","tarkistusarvioinninAsiatunnus","tarkistusarvioidutOsakokeet",arvosanaMuuttui,perustelu,"tarkistusarvioinninKasittelyPvm"
@@ -144,7 +153,8 @@ class CsvParsingTest {
     fun `null values are written correctly to csv`() {
         val datePattern = "yyyy-MM-dd"
         val dateFormatter = DateTimeFormatter.ofPattern(datePattern)
-        val event = LoggerFactory.getLogger(CsvArgs::class.java).atInfo()
+        val event = LoggerFactory.getLogger(CsvParsingTest::class.java).atInfo()
+        val parser = CsvParser(event, useHeader = true)
         val entity =
             YkiSuoritusEntity(
                 id = null,
@@ -181,7 +191,8 @@ class CsvParsingTest {
             )
         val writable = listOf(entity.toYkiSuoritusCsv())
         val outputStream = ByteArrayOutputStream()
-        writable.writeAsCsv(outputStream, CsvArgs(useHeader = true, event = event))
+        parser.streamDataAsCsv(outputStream, writable)
+
         val expectedCsv =
             """
             suorittajanOID,hetu,sukupuoli,sukunimi,etunimet,kansalaisuus,katuosoite,postinumero,postitoimipaikka,email,suoritusID,lastModified,tutkintopaiva,tutkintokieli,tutkintotaso,jarjestajanOID,jarjestajanNimi,arviointipaiva,tekstinYmmartaminen,kirjoittaminen,rakenteetJaSanasto,puheenYmmartaminen,puhuminen,yleisarvosana,"tarkistusarvioinninSaapumisPvm","tarkistusarvioinninAsiatunnus","tarkistusarvioidutOsakokeet",arvosanaMuuttui,perustelu,"tarkistusarvioinninKasittelyPvm"
